cleanup installer, use temporary installation folder

+++ src/Downloader/OctoberCms.php
@@ -14,10 +14,23 @@ use ZipArchive;
 
 class OctoberCms extends BaseManager
 {
-    protected $zipFile;
+    protected $files = [
+        'bootstrap' => '',
+        'config' => '',
+        'modules' => '',
+        'plugins' => '',
+        'storage' => '',
+        'tests' => '',
+        'themes' => '',
+        '.htaccess' => '.htaccess',
+        'artisan' => 'artisan',
+        'composer.json' => 'composer.json',
+        'index.php' => 'index.php',
+        'server.php' => 'server.php',
+    ];
+
     protected $installerFile;
-    protected $htaccessFileSrc = "https://raw.githubusercontent.com/octobercms/october/1.1/.htaccess";
-    protected $ocmsInstallDir = '.ocms';
+    protected $ocmsInstallDir = '/tmp/ocms';
 
     /**
      * Downloads and extracts October CMS.
@@ -26,7 +39,6 @@ class OctoberCms extends BaseManager
     public function __construct()
     {
         parent::__construct();
-        $this->zipFile = $this->makeFilename();
         $this->installerFile = getcwd() . DS . 'installer.php';
     }
 
@@ -41,56 +53,28 @@ class OctoberCms extends BaseManager
      */
     public function download($force = false)
     {
+        $this->write('-> Checking if October CMS download is required...');
+
         if ($this->alreadyInstalled($force)) {
             throw new \LogicException('-> October is already installed. Use --force to reinstall.');
         }
 
-
-        /*
-        $this->fetchZip()
-             ->extract()
-             ->fetchHtaccess()
-             ->cleanUp()
-             ->setMaster();
-        */
-
-        /*
-        $this->downloadInstaller()
-            ->install()
-            ->fetchHtaccess()
-            ->cleanUpInstaller();
-        */
-        $this->composerInstaller();
-        return $this;
-    }
-
-    protected function downloadInstaller()
-    {
-        $url = 'https://octobercms.com/api/installer';
-        $this->write("-> Downloading $url");
-        $response = (new Client)->get($url);
-        file_put_contents($this->installerFile, $response->getBody());
-
-
-        return $this;
-    }
-
-    protected function composerInstaller()
-    {
         $this
             ->cleanupProject()
             ->createProject()
             ->copyProjectFiles()
             ->cleanupProject();
 
-
         return $this;
     }
 
     protected function cleanupProject()
     {
-        $this->write("-> Deleting ocms copy in ".$this->ocmsInstallDir);
-        (new Process(sprintf('rm -rf %s', $this->ocmsInstallDir)))->run();
+        if (is_dir($this->ocmsInstallDir)) {
+            $this->write("-> Deleting ocms copy in '$this->ocmsInstallDir'");
+            (new Process(sprintf('rm -rf %s', $this->ocmsInstallDir)))->run();
+        }
+
         return $this;
     }
 
@@ -104,25 +88,9 @@ class OctoberCms extends BaseManager
 
     protected function copyProjectFiles()
     {
-        $files = [
-            'bootstrap',
-            'config',
-            'modules',
-            'plugins',
-            'storage',
-            'tests',
-            'themes',
-            '.htaccess',
-            'artisan',
-            '.gitignore',
-            'composer.json',
-            'index.php',
-            'server.php'
-        ];
-
-        foreach ($files as  $file) {
-            $src = $this->ocmsInstallDir . DS . $file;
-            $dst = $file;
+        foreach ($this->files as  $src => $dst) {
+            $src = $this->ocmsInstallDir . DS . $src;
+            $dst = $this->pwd() . $dst;
             $this->write("-> copying ".$src." -> ".$dst);
 
             (new Process(sprintf('cp -rn %s %s', $src, $dst)))->run();
@@ -131,73 +99,6 @@ class OctoberCms extends BaseManager
         return $this;
 
     }
-
-    protected function install()
-    {
-        $this->write("-> Execute installer 'php " . $this->installerFile ."'");
-        (new Process(sprintf($this->php.' -f %s', $this->installerFile)))->run();
-
-        return $this;
-    }
-
-    protected function cleanUpInstaller()
-    {
-        $this->write("-> Cleanup installation file ". $this->installerFile);
-        (new Process(sprintf('rm -f %s', $this->installerFile)))->run();
-        return $this;
-    }
-
-    /**
-     * Download the temporary Zip to the given file.
-     *
-     * @return $this
-     * @throws RuntimeException
-     * @throws LogicException
-     */
-    protected function fetchZip()
-    {
-        $url = 'https://github.com/octobercms/october/archive/master.zip';
-        $this->write("-> Downloading $url");
-        $response = (new Client)->get($url);
-        file_put_contents($this->zipFile, $response->getBody());
-
-        return $this;
-    }
-
-    /**
-     * Extract the zip file into the given directory.
-     *
-     * @return $this
-     */
-    protected function extract()
-    {
-
-        $dst = getcwd();
-        $this->write("-> Extracting zip to $dst");
-
-        $archive = new ZipArchive;
-        $archive->open($this->zipFile);
-        $archive->extractTo($dst);
-        $archive->close();
-
-        return $this;
-    }
-
-    /**
-     * Download the latest .htaccess file from GitHub separately
-     * since ZipArchive does not support extracting hidden files.
-     *
-     * @return $this
-     */
-    protected function fetchHtaccess()
-    {
-        $this->write("-> Downloading .htaccess file from ".$this->htaccessFileSrc);
-        $contents = file_get_contents($this->htaccessFileSrc);
-        file_put_contents(getcwd() . DS . '.htaccess', $contents);
-
-        return $this;
-    }
-
 
     /**
      * Since we don't want any unstable updates we fix
@@ -210,7 +111,6 @@ class OctoberCms extends BaseManager
         $json = getcwd() . DS . 'composer.json';
 
         $this->write("-> Changing October CMS dependencies to dev-master");
-
 
         $contents = file_get_contents($json);
 
@@ -230,48 +130,23 @@ class OctoberCms extends BaseManager
     }
 
     /**
-     * Remove the Zip file, move folder contents one level up.
-     *
-     * @return $this
-     * @throws RuntimeException
-     * @throws LogicException
-     */
-    protected function cleanUp()
-    {
-        @chmod($this->zipFile, 0777);
-        @unlink($this->zipFile);
-
-        $directory = getcwd();
-        $source    = $directory . DS . 'october-master';
-
-        (new Process(sprintf('mv %s %s', $source . '/*', $directory)))->run();
-        (new Process(sprintf('rm -rf %s', $source)))->run();
-
-        if (is_dir($source)) {
-            echo "<comment>Install directory could not be removed. Delete ${source} manually</comment>";
-        }
-
-        return $this;
-    }
-
-    /**
-     * Generate a random temporary filename.
-     *
-     * @return string
-     */
-    protected function makeFilename()
-    {
-        return getcwd() . DS . 'october_' . md5(time() . uniqid('oc-', true)) . '.zip';
-    }
-
-    /**
      * @param $force
      *
      * @return bool
      */
     protected function alreadyInstalled($force)
     {
-        return ! $force && is_dir(getcwd() . DS . 'bootstrap') && is_dir(getcwd() . DS . 'modules');
+        if ($force) return false;
+
+        foreach ($this->files as $file => $target) {
+            $realFile = getcwd() . DS . $file;
+            if (!is_dir($realFile) && !is_file($realFile)) {
+                $this->write("-> Missing file or dir '/$file'");
+                return false;
+            }
+        }
+
+        return true;
     }
 
 }
